Reorganize texting functionality
Break out the texting functionality from `views.py` into its own
separate file. Additionally, provide an example of how tests can work.
