+++ text_support/text.py
@@ -1,0 +1,17 @@
+"""
+`text.py` is where we contain classes and functions related to texting with the
+user.
+"""
+
+def get_response(message):
+    """
+    Given what the texter has sent to us, determine our response.
+
+    Args:
+        message (str): The message the texter has sent to us.
+
+    Response:
+        str: The response we will text back to them.
+    """
+    # @TODO Fill this function in with code that will have respond.
+    return message
